move preseed utils to seed folder, add unit testing structure
